Allow autocomplete for testing image paths
Also - stop deleting the files!

+++ digits/model/images/classification/views.py
@@ -288,15 +288,17 @@ def classify_one():
     """
     model_job = job_from_request()
 
-    if 'image_url' in flask.request.form and flask.request.form['image_url']:
-        image_path = flask.request.form['image_url']
+    remove_image_path = False
+    if 'image_path' in flask.request.form and flask.request.form['image_path']:
+        image_path = flask.request.form['image_path']
     elif 'image_file' in flask.request.files and flask.request.files['image_file']:
         outfile = tempfile.mkstemp(suffix='.png')
         flask.request.files['image_file'].save(outfile[1])
         image_path = outfile[1]
         os.close(outfile[0])
+        remove_image_path = True
     else:
-        raise werkzeug.exceptions.BadRequest('must provide image_url or image_file')
+        raise werkzeug.exceptions.BadRequest('must provide image_path or image_file')
 
     epoch = None
     if 'snapshot_epoch' in flask.request.form:
@@ -328,11 +330,8 @@ def classify_one():
     # delete job
     scheduler.delete_job(inference_job)
 
-    # remove file (fails silently if a URL was provided)
-    try:
+    if remove_image_path:
         os.remove(image_path)
-    except:
-        pass
 
     image = None
     predictions = []
